# web/swagger/swagger.yaml
consumes:
- application/x-www-form-urlencoded
definitions:
  CreateRecipePayload:
    example:
      complete: false
      cook_time: 2003-08-13T12:47:16Z
      description: Est facilis commodi ratione temporibus.
      difficulty: 0.52087873
      favorite: false
      image: Doloremque qui incidunt autem.
      images:
      - Occaecati totam deleniti iusto quos.
      ingredients:
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      prep_time: 1990-08-15T00:54:39Z
      rating: 0.9356256
      source: Recusandae aut.
      state: Laboriosam enim accusamus in libero.
      title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
      version: Consequuntur occaecati ipsa qui.
      wait_time: 2004-11-23T22:16:37Z
    properties:
      complete:
        description: If it's been added/included
        example: false
        type: boolean
      cook_time:
        description: Amount of time to cook
        example: 2003-08-13T12:47:16Z
        format: date-time
        type: string
      description:
        description: Long description of recipe
        example: Est facilis commodi ratione temporibus.
        type: string
      difficulty:
        description: rating between 0-1
        example: 0.52087873
        maximum: 1
        minimum: 0
        type: number
      favorite:
        description: Is a favorite, basically a tag
        example: false
        type: boolean
      image:
        description: Image of recipe
        example: Doloremque qui incidunt autem.
        type: string
      images:
        description: Images of recipe
        example:
        - Occaecati totam deleniti iusto quos.
        items:
          example: Occaecati totam deleniti iusto quos.
          type: string
        type: array
      ingredients:
        description: List of ingredients
        example:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        items:
          $ref: '#/definitions/RecipePayload'
        type: array
      prep_time:
        description: Amount of time to prepare
        example: 1990-08-15T00:54:39Z
        format: date-time
        type: string
      rating:
        description: rating between 0-1
        example: 0.9356256
        maximum: 1
        minimum: 0
        type: number
      source:
        description: Source of recipe
        example: Recusandae aut.
        type: string
      state:
        description: e.g. chopped, sliced, etc.. might need to be array.
        example: Laboriosam enim accusamus in libero.
        type: string
      title:
        description: Recipe Title
        example: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        type: string
      version:
        description: Version Number e.g. 1.0.1
        example: Consequuntur occaecati ipsa qui.
        type: string
      wait_time:
        description: Amount of time to wait for things such as mairnading
        example: 2004-11-23T22:16:37Z
        format: date-time
        type: string
    required:
    - title
    title: CreateRecipePayload
    type: object
...
  UpdateRecipePayload:
    example:
      complete: false
      cook_time: 2003-08-13T12:47:16Z
      description: Est facilis commodi ratione temporibus.
      difficulty: 0.52087873
      favorite: false
      image: Doloremque qui incidunt autem.
      images:
      - Occaecati totam deleniti iusto quos.
      ingredients:
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      - complete: false
        cook_time: 2003-08-13T12:47:16Z
        description: Est facilis commodi ratione temporibus.
        difficulty: 0.52087873
        favorite: false
        image: Doloremque qui incidunt autem.
        images:
        - Occaecati totam deleniti iusto quos.
        ingredients:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        prep_time: 1990-08-15T00:54:39Z
        rating: 0.9356256
        source: Recusandae aut.
        state: Laboriosam enim accusamus in libero.
        title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        version: Consequuntur occaecati ipsa qui.
        wait_time: 2004-11-23T22:16:37Z
      prep_time: 1990-08-15T00:54:39Z
      rating: 0.9356256
      source: Recusandae aut.
      state: Laboriosam enim accusamus in libero.
      title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
      version: Consequuntur occaecati ipsa qui.
      wait_time: 2004-11-23T22:16:37Z
    properties:
      complete:
        description: If it's been added/included
        example: false
        type: boolean
      cook_time:
        description: Amount of time to cook
        example: 2003-08-13T12:47:16Z
        format: date-time
        type: string
      description:
        description: Long description of recipe
        example: Est facilis commodi ratione temporibus.
        type: string
      difficulty:
        description: rating between 0-1
        example: 0.52087873
        maximum: 1
        minimum: 0
        type: number
      favorite:
        description: Is a favorite, basically a tag
        example: false
        type: boolean
      image:
        description: Image of recipe
        example: Doloremque qui incidunt autem.
        type: string
      images:
        description: Images of recipe
        example:
        - Occaecati totam deleniti iusto quos.
        items:
          example: Occaecati totam deleniti iusto quos.
          type: string
        type: array
      ingredients:
        description: List of ingredients
        example:
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        - complete: false
          cook_time: 2003-08-13T12:47:16Z
          description: Est facilis commodi ratione temporibus.
          difficulty: 0.52087873
          favorite: false
          image: Doloremque qui incidunt autem.
          images:
          - Occaecati totam deleniti iusto quos.
          ingredients:
          - complete: false
            cook_time: 2003-08-13T12:47:16Z
            description: Est facilis commodi ratione temporibus.
            difficulty: 0.52087873
            favorite: false
            image: Doloremque qui incidunt autem.
            images:
            - Occaecati totam deleniti iusto quos.
            prep_time: 1990-08-15T00:54:39Z
            rating: 0.9356256
            source: Recusandae aut.
            state: Laboriosam enim accusamus in libero.
            title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
            version: Consequuntur occaecati ipsa qui.
            wait_time: 2004-11-23T22:16:37Z
          prep_time: 1990-08-15T00:54:39Z
          rating: 0.9356256
          source: Recusandae aut.
          state: Laboriosam enim accusamus in libero.
          title: Itaque maiores quasi corrupti rerum aliquid reiciendis.
          version: Consequuntur occaecati ipsa qui.
          wait_time: 2004-11-23T22:16:37Z
        items:
          $ref: '#/definitions/RecipePayload'
        type: array
      prep_time:
        description: Amount of time to prepare
        example: 1990-08-15T00:54:39Z
        format: date-time
        type: string
      rating:
        description: rating between 0-1
        example: 0.9356256
        maximum: 1
        minimum: 0
        type: number
      source:
        description: Source of recipe
        example: Recusandae aut.
        type: string
      state:
        description: e.g. chopped, sliced, etc.. might need to be array.
        example: Laboriosam enim accusamus in libero.
        type: string
      title:
        description: Recipe Title
        example: Itaque maiores quasi corrupti rerum aliquid reiciendis.
        type: string
      version:
        description: Version Number e.g. 1.0.1
        example: Consequuntur occaecati ipsa qui.
        type: string
      wait_time:
        description: Amount of time to wait for things such as mairnading
        example: 2004-11-23T22:16:37Z
        format: date-time
        type: string
    required:
    - title
    title: UpdateRecipePayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
...
host: www.recipe.com
info:
  contact:
    email: jlwarren1@gmail.com
    name: Jared Warren
    url: http://jlwarren1.com
  description: CRUD API recipe
  license:
    name: MIT
    url: https://github.com/goadesign/goa/blob/master/LICENSE
  title: Recipe Web interrace
  version: ""
paths:
...
  /recipe/:
    get:
      description: List recipes
      operationId: recipe#list
      produces:
      - application/vnd.goa.error
      - text/html
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: list recipe
      tags:
      - recipe
    post:
      operationId: recipe#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateRecipePayload'
      produces:
      - application/recipe.recipe+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RecipeRecipe'
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: create recipe
      tags:
      - recipe
  /recipe/{id}:
    delete:
      operationId: recipe#delete
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: delete recipe
      tags:
      - recipe
    get:
      description: Display an recipe by id
      operationId: recipe#show
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - text/html
      responses:
        "200":
          description: OK
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: text/html
              type: string
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: show recipe
      tags:
      - recipe
    patch:
      operationId: recipe#update
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateRecipePayload'
      produces:
      - application/recipe.recipe+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RecipeRecipe'
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: update recipe
      tags:
      - recipe
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
swagger: "2.0"